Reduce the occupied memory of the Listpack object.
Adds a way to use the tagged pointers to be able to get
a view onto a listpack from a borrowed pointer.

A better way for the future is to use strict-typed structs
which would discriminate between the borrowed and owned pointers used
for the listpack allocation easier and in a safer way.